requirements.txt
asgiref==3.8.1
coverage==7.9.1
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-graphene-auth==1.1.1
django-graphql-jwt==0.4.0
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.4.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
promise==2.3
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.0
pytest-cov==6.2.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.0
tzdata==2025.2
